soft delete v1

+++ app/Models/File.php
@@ -1,18 +1,22 @@
 <?php
-
 namespace App\Models;
 
 use Illuminate\Support\Str;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Database\Eloquent\SoftDeletes; // Import SoftDeletes trait
 
 class File extends Model
 {
+    use SoftDeletes; // Use the trait
+
     protected $fillable = [
         'name',
         'size',
         'path'
     ];
+
+    protected $dates = ['deleted_at']; // Add the deleted_at column
 
     public function sizeForHumans()
     {
@@ -33,7 +37,12 @@ class File extends Model
             $model->uuid = Str::uuid();
         });
 
+        // Modify the deleting event to prevent physical deletion
         static::deleting(function ($model) {
+            if (!$model->isForceDeleting()) {
+                return;
+            }
+
             Storage::disk('local')->delete($model->path);
         });
     }
